functions/check-subscription-status/requirements.txt
ydb==3.8.1
PyJWT==2.8.0